Unify particle editors
